pyproject.toml
[project]
name = "python-game-vault"
version = "0.1.0"
description = "Claude-powered backend for managing an Obsidian RPG campaign vault"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-frontmatter>=1.1.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
